move fuzzy find something to nucleo picker
